241 override property
