+++ app/build.gradle
@@ -25,6 +25,7 @@ android {
 
         buildConfigField "String", "youtubeApiKey", "\"$CUER_YOUTUBE_API_KEY\""
         buildConfigField "String", "pixabayApiKey", "\"$CUER_PIXABAY_API_KEY\""
+        buildConfigField "boolean", "backgroundPlay", "$CUER_BG_PLAY"
 
         javaCompileOptions {
             annotationProcessorOptions {
