+++ app/src/main/java/uk/co/sentinelweb/cuer/app/ui/ytplayer/floating/FloatingPlayerController.kt
@@ -1,6 +1,7 @@
 package uk.co.sentinelweb.cuer.app.ui.ytplayer.floating
 
 import android.content.Intent
+import uk.co.sentinelweb.cuer.app.BuildConfig
 import uk.co.sentinelweb.cuer.app.receiver.ScreenStateReceiver
 import uk.co.sentinelweb.cuer.app.service.cast.notification.player.PlayerControlsNotificationController.Companion.ACTION_DISCONNECT
 import uk.co.sentinelweb.cuer.app.service.cast.notification.player.PlayerControlsNotificationController.Companion.ACTION_PAUSE
@@ -17,6 +18,7 @@ import uk.co.sentinelweb.cuer.app.ui.player.PlayerController
 import uk.co.sentinelweb.cuer.app.ui.ytplayer.AytViewHolder
 import uk.co.sentinelweb.cuer.app.ui.ytplayer.floating.FloatingPlayerService.Companion.ACTION_INIT
 import uk.co.sentinelweb.cuer.app.ui.ytplayer.floating.FloatingPlayerService.Companion.ACTION_PLAY_ITEM
+import uk.co.sentinelweb.cuer.app.util.prefs.multiplatfom_settings.MultiPlatformPreferencesWrapper
 import uk.co.sentinelweb.cuer.app.util.wrapper.ToastWrapper
 import uk.co.sentinelweb.cuer.core.wrapper.LogWrapper
 import uk.co.sentinelweb.cuer.domain.ext.deserialisePlaylistItem
@@ -29,7 +31,8 @@ class FloatingPlayerController constructor(
     private val aytViewHolder: AytViewHolder,
     private val log: LogWrapper,
     private val screenStateReceiver: ScreenStateReceiver,
-    private val toastWrapper: ToastWrapper
+    private val toastWrapper: ToastWrapper,
+    private val multiPrefs: MultiPlatformPreferencesWrapper,
 ) : FloatingPlayerContract.Controller, FloatingPlayerContract.External {
 
     init {
@@ -45,10 +48,32 @@ class FloatingPlayerController constructor(
             playerMviViw.mainPlayControls = value
         }
 
+    private var wasPausedScreenLocked: Boolean = false
+
+    private val lockHandler = {
+        log.d("Screen off -> ACTION_PAUSE")
+        playerMviViw.setTitlePrefix("[locked]")
+        handleAction(Intent(ACTION_PAUSE))
+        wasPausedScreenLocked = true
+    }
+
+    private val unlockHandler = {
+        log.d("Unlock -> ACTION_PLAY")
+        playerMviViw.setTitlePrefix(null)
+        if (wasPausedScreenLocked && multiPrefs.restartAfterUnlock) {
+            handleAction(Intent(ACTION_PLAY))
+        }
+        wasPausedScreenLocked = false
+    }
+
     override fun initialise() {
         windowManagement.makeWindowWithView()
         windowManagement.callbacks = object : FloatingWindowManagement.Callbacks {
             override fun onClose() = service.stopSelf()
+        }
+        if (!BuildConfig.backgroundPlay) {
+            screenStateReceiver.screenOffCallbacks.add(lockHandler)
+            screenStateReceiver.unlockCallbacks.add(unlockHandler)
         }
         playerMviViw.init()
         playerController.onViewCreated(listOf(playerMviViw))
@@ -56,6 +81,10 @@ class FloatingPlayerController constructor(
     }
 
     override fun destroy() {
+        if (!BuildConfig.backgroundPlay) {
+            screenStateReceiver.screenOffCallbacks.remove(lockHandler)
+            screenStateReceiver.unlockCallbacks.remove(unlockHandler)
+        }
         playerMviViw.cleanup()
         playerController.onStop()
         playerController.onViewDestroyed()
@@ -70,6 +99,7 @@ class FloatingPlayerController constructor(
             ACTION_DISCONNECT -> {
                 service.stopSelf()
             }
+
             ACTION_STAR -> Unit
             // log.d(serviceWrapper.getServiceData(YoutubeCastService::class.java.name).toString())
             ACTION_INIT -> {
@@ -78,6 +108,7 @@ class FloatingPlayerController constructor(
                     ?.let { deserialisePlaylistItem(it) }
                     ?.also { playerMviViw.dispatch(OnInitFromService(it)) }
             }
+
             ACTION_PLAY_ITEM -> {
                 intent
                     .getStringExtra(NavigationModel.Param.PLAYLIST_ITEM.toString())
@@ -95,21 +126,21 @@ class FloatingPlayerController constructor(
                 playerMviViw.dispatch(PlayPauseClicked(true))
 
             ACTION_PLAY ->
-                if (!screenStateReceiver.isLocked) {
+                if (allowPlayInBackground()) {
                     playerMviViw.dispatch(PlayPauseClicked(false))
                 } else {
                     toastWrapper.show("Unlock to play")
                 }
 
             ACTION_TRACKB ->
-                if (!screenStateReceiver.isLocked) {
+                if (allowPlayInBackground()) {
                     playerMviViw.dispatch(TrackBackClicked)
                 } else {
                     toastWrapper.show("Unlock to play")
                 }
 
             ACTION_TRACKF ->
-                if (!screenStateReceiver.isLocked) {
+                if (allowPlayInBackground()) {
                     playerMviViw.dispatch(TrackFwdClicked)
                 } else {
                     toastWrapper.show("Unlock to play")
@@ -121,7 +152,5 @@ class FloatingPlayerController constructor(
         }
     }
 
-    override fun setTitlePrefix(prefix: String?) {
-        playerMviViw.setTitlePrefix(prefix)
-    }
+    private fun allowPlayInBackground() = !screenStateReceiver.isLocked || BuildConfig.backgroundPlay
 }
