+++ gradle.properties
@@ -24,9 +24,11 @@ isProduction=false
 # jdk
 ver_jdk=8
 ver_jvm=1.8
-# android vars
+# release vars
 app_versionCode=6
 app_versionName=0.74-beta
+CUER_BG_PLAY=true
+# android vars
 app_applicationId=uk.co.sentinelweb.cuer.app
 app_compileSdkVersion=32
 app_targetSdkVersion=30
